hoàn thiện nốt phần phân quyền

+++ src/main/java/com/example/demo/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.example.demo.controller;
 
-import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -13,8 +12,14 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.example.demo.dto.UserRequest;
 import com.example.demo.dto.UserResponse;
+import com.example.demo.exception.BadRequestException;
+import com.example.demo.exception.NotFoundException;
+import com.example.demo.model.User;
+import com.example.demo.repository.UserRepository;
+import com.example.demo.security.JwtUtil;
 import com.example.demo.service.UserService;
 
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 
 
@@ -25,10 +30,11 @@ public class UserController {
 	@Autowired
 	private UserService userService;
 	
-	@GetMapping
-	public List<UserResponse> getAllUser(){
-		return userService.getAllUser();
-	}
+	@Autowired
+	private JwtUtil jwtUtil;
+	
+	@Autowired
+	private UserRepository userRepository;
 	
 	@GetMapping("/{id}")
 	public UserResponse getUserById(@PathVariable Long id) {
@@ -36,12 +42,38 @@ public class UserController {
 	}
 	
 	@PutMapping("/update/{id}")
-	public UserResponse updateUser(@PathVariable Long id,@Valid @RequestBody UserRequest userRequest) {
+	public UserResponse updateUser(@PathVariable Long id,@Valid @RequestBody UserRequest userRequest,HttpServletRequest request) {
+		String authHeader = request.getHeader("Authorization");
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            throw new BadRequestException("Token không hợp lệ");
+        }
+
+        String token = authHeader.substring(7);
+        String email = jwtUtil.extractEmail(token);
+
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new NotFoundException("User not found"));
+        if(user.getId()!=id) {
+            return null;
+        }
 		return userService.updateUser(id, userRequest);
 	}
 	
 	@DeleteMapping("/delete/{id}")
-	public String deleteUser(@PathVariable Long id) {
+	public String deleteUser(@PathVariable Long id,HttpServletRequest request) {
+		String authHeader = request.getHeader("Authorization");
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            throw new BadRequestException("Token không hợp lệ");
+        }
+
+        String token = authHeader.substring(7);
+        String email = jwtUtil.extractEmail(token);
+
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new NotFoundException("User not found"));
+        if(user.getId()!=id) {
+        	return "Bạn không có quền xóa tài khoản này";
+        }
 		userService.deleteUser(id);
 		return "User deleted successfully";
 	}
